+++ src/BubbleSort.h
@@ -1,4 +1,4 @@
 namespace Sort::Bubble {
 	void Apply(int *data, int n);
-	int CountComparisons(int *data, int n);
+	long long CountComparisons(int *data, int n);
 }
